rename: axiosInstance 네이밍 변경 후 폴더이동
