+++ src/api/plan.ts
@@ -4,7 +4,7 @@ import {
   CreatePlanResponse,
   PlanDetailResponse,
 } from '@/types/api/plan';
-import instance from './axiosInstance';
+import instance from '../utils/axios';
 import { ApiResponse } from '@/types/api/apiResponse';
 import { showToast } from '@/utils/showToast';
 import TOAST_MESSAGE from '@/constants/toastMessage';
