+++ src/api/reviews.test.ts
@@ -1,5 +1,5 @@
 import fetchReviews from '@/api/reviews';
-import axiosInstance from '@/api/axiosInstance';
+import axiosInstance from '@/utils/axios';
 import { FilterState } from '@/types/reviewType';
 
 jest.mock('@/api/axiosInstance'); // axiosInstance를 목 함수로 만들어줍니다.
