+++ src/pages/plans/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import type { NextPage, GetServerSideProps } from 'next';
 //import axios from 'axios';
-import instance from '@/api/axiosInstance';
+import instance from '@/utils/axios';
 import { SortOption } from '@/types/reviewType';
 import { useCursorInfiniteScroll } from '@/hooks/useCursorInfiniteScroll';
 import { PlanDataWithCategory } from '@/types/plans';
